Fix extension check for xfb mode validation
With geometry and tessellation shaders, it's possible to use different
primitive modes between transform feedback and draw, as long as they are
from the same class.

The code accidentally allowed this only if both extensions are
supported, but it should be allowed if either is.

Bug: angleproject:428561247
Change-Id: Ia18d4ac15a3d0739ce18dc0e1f3d3f0b1aea621c
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6780744
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>