tests: RequestedRendererCreated: Get created context version
If the driver being tested does not support the ANGLE extension
EGL_ANGLE_create_context_backwards_compatible, then the Context version
returned may not match the requested version.

Update the test RendererTest.RequestedRendererCreated to query the
current Context's version if the extension is not supported so the
renderer string is checked against the correct value. Continue enforcing
the renderer string has the expected value if the extension is
supported.

Bug: b/409384875
Test: angle_end2end_tests
  --gtest_filter=RendererTest.RequestedRendererCreated/*
Change-Id: Iad1b6d36a31b7d79e8fc5fae9e4cc9769b55edb8
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6803326
Commit-Queue: Tim Van Patten <timvp@google.com>
Reviewed-by: Yuxin Hu <yuxinhu@google.com>
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>