Use the chrome plugin in select Angle build recipes
Turn on the chrome plugin for use in future CLs so that additional
checks may be someday performed against this code base.

This CL should not affect any generated code or produce additional
warnings. To do so, remove find_bad_constructs as code was not
compliant, just not subject to the plugin.

Bug: b/436880895
Change-Id: I73bcc263fe0fe1656bce5a75a74e270ecf004362
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6825145
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>