TestSuite: Android: Call onCrashOrTimeout() on test crash
Currently, when a test crashes on Android we don't get any results for
tests that ran or for the crashing test itself. This makes it difficult
for developers to debug the issue, since they don't know which test
crashed.

Add signal handlers which call onCrashOrTimeout() when a test crashes.
Typically, crashes occur within the individual tests rather than in the
test runner (TestSuite), so it should be "safe" to attempt to dump the
results up to and including the crash to aid with debugging.

This is still risky since things are in a bad state, but we expect
(essentially) all crashes to be in the individual tests, ANGLE, or the
underlying driver code, rather than in the TestSuite instance, so
writing the results JSON file should have a good chance of succeeding.
Worst case, we just crash again and get nothing (like we have before
this change).

Bug: b/279980674
Change-Id: Ifd5f3b97d5af731586e7bbf765bb60b3cb4f503b
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6827986
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Commit-Queue: Tim Van Patten <timvp@google.com>
Reviewed-by: Cody Northrop <cnorthrop@google.com>
Reviewed-by: Solti Ho <solti@google.com>