Metal: Fix potential incorrect format used for texSubImage
TextureMtl caches the most recent format used for a texture
redefinition in mFormat. During subImage calls, this format may not be
the same as the image being uploaded to.

Pass ImageDefinitionMtl to these functions which reference mFormat, it
contains the format of the image being updated.

MacOS OpenGL is skipped because the driver generates errors in
glCompressedTexImage2D.

Bug: chromium:435683799
Change-Id: Idec6f71870c2d376cad3a5e3628b957009bdced9
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6838878
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Reviewed-by: Quyen Le <lehoangquyen@chromium.org>
Commit-Queue: Geoff Lang <geofflang@chromium.org>