Vulkan: Check format compatibility to enable mip levels
If redefining the texture level is needed while respecifying the
image storage, the base level could be incompatible with the max
level count.
Enabling the whole level count can cause a null pointer dereference
when copying an image that does not support `pixelReadFunction`.
To handle the problem, this CL checks the format compatibility and
enables only the effective base level.

Bug: b/435727832
Test: angle_end2end_tests --gtest_filter=DXT1CompressedTextureTest.*
Test: angle_end2end_tests --gtest_filter=Texture2DTestES3.*
Change-Id: I01060332f282ef9b2856390d0d7296811a464795
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6813835
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>