Update setFloat and getFloat APIs to handle 16-32 bit transformations
This change checks that if linkedUniform.isFloat16() is true, then
transform 32-bit float to 16-bit half float before writing the data
to memory, and transform 16-bit half float to 32-bit float after
reading the data from memory. Given that we did not change buffer
layout, and the spirv requirement with matrix component alignment,
we have to copy transformed data row by row / col by col, or element
by element.

This change is no op, as none of the linkedUniform.isFloat() bit is set
to true. The change will take effect in the future change when we add
the compiler code that set linkedUniform.isFloat() to true.

Bug: angleproject:405795981
Change-Id: I72b74eb340dae723501038875c3706ad2c1cc3f7
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6851558
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Commit-Queue: Yuxin Hu <yuxinhu@google.com>
Reviewed-by: Amirali Abdolrashidi <abdolrashidi@google.com>