Reorder structs entirely made of samplers to the end of struct
An earlier change [1] moved samplers to the end of structs so that
extracting them does not disturb the indices to the non-sampler fields.
This change does the same for structs that only contain samplers,
because they too get removed from the parent struct, disturbing indices.

[1]:https://chromium-review.googlesource.com/c/angle/angle/+/5748789

Bug: chromium:438038775
Change-Id: I7ac2ea13723bf06b4b172b6e70f0dbe85bc45ce3
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6891071
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>
Reviewed-by: Geoff Lang <geofflang@chromium.org>