GL: Disable BaseVertex and BaseInstance extensions on Linux
This is a speculative fix for some rendering issues seen on Linux.

Client data also appears to be broken when using BaseVertex functions,
this will disable the functions until those issues are fixed.

Re-add the check for GL_ARB_base_instance, this was likely not the
culprit.

Bug: chromium:427956856, angleproject:431097618
Change-Id: If48121d3c4819a46091de4360a2159b9bf14313e
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6897307
Commit-Queue: Geoff Lang <geofflang@chromium.org>
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>