[WebGPU] Fix ImageHelper::flushSingleLevelUpdates
This change modifies ImageHelper:flushSingleLevelUpdates to
only send a flush command to the context once, instead of in each
iteration of the loop if there are multiple texture uploads.

This also means that the last texture upload is guaranteed to be
flushed, as otherwise that texture could be staged in the queue,
but a different command could redefine texture levels which
would delete the underlying texture handle but not the command
queued for that texture.

Bug: angleproject:438268609
Change-Id: I2d49b15fc921829175e99bdb550dddbd81179061
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6885223
Commit-Queue: Matthew Denton <mpdenton@chromium.org>
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Reviewed-by: Matthew Denton <mpdenton@chromium.org>