Don't use DXGI_SWAP_CHAIN_FLAG_FRAME_LATENCY_WAITABLE_OBJECT
A swap chain for Windows.UI.Composition is created with
DXGI_SWAP_CHAIN_FLAG_FRAME_LATENCY_WAITABLE_OBJECT flag,
but is resized without it. This leads to an error,
because according to MSDN, this flag cannot be changed
by a call to ResizeBuffers.

In fact, this flag is never used by any code in ANGLE,
so we should remove it.

Bug: angleproject:442600654
Change-Id: I3eecd96abd243433b402780c30634abf2528c323
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6919163
Commit-Queue: Geoff Lang <geofflang@chromium.org>
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Reviewed-by: Geoff Lang <geofflang@chromium.org>