Add condition check in emulate dithering shader code
Prior to this change, the emulate dithering shader code will write to
all elements of gl_FragData, regardless of whether the dither value is
zero or not: gl_FragData[i] = gl_FragData[i] + dither_value.
If dither_value is 0, then this doesn't change gl_FragData[i] and it
incurs unnecessary shader writes.

This change adds a boolean variable that is initialized to
false, and it is only set to true if dither value for that output
element is changed from zero to non-zero value. This way we are only
issuing shader write operations when necessary.

Bug: angleproject:425733272
Change-Id: Iacfd81ff9a1d7d0f71ecf8d119033b2500d2869e
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6931777
Reviewed-by: Charlie Lao <cclao@google.com>
Commit-Queue: Yuxin Hu <yuxinhu@google.com>
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>