HLSL: Mark shadow sampler textureLod calls as using gradients
TextureLod calls bottom out in the HLSL SampleCmp function which does
not take a LOD parameter so we end up dropping it. This causes them to
need gradients. Mark them as using gradient so that any loop they are
used in is flagged and not unrolled by FXC.

Reverted original fix from
https://chromium-review.googlesource.com/c/angle/angle/+/6421496
because the wrong number of parameters are passed to the generated
comparison sampling function when it is changed to LOD0.

Thanks to Mikkel Simonsen at Unity and Calder Young for the tests.

Fixed: angleproject:406190894
Change-Id: Ia9255e4336099045366c02f945a3c4de8075427a
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6936488
Commit-Queue: Geoff Lang <geofflang@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>