Vulkan: Split handleDirtyGraphicsVertexBuffers into two funcs
ContextVk::handleDirtyGraphicsVertexBuffers() has two code paths that
are quite different, depends on if supportsVertexInputDynamicState is
enabled or not. This CL splits this function into two so that 1) code is
much easier to read that what is done in each case and 2) get rid of if
check at draw time.

Bug: b/439073246
Change-Id: Iaf9f84fb20d82e784f9dfe8ece321cff94b7b58d
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6956208
Commit-Queue: Charlie Lao <cclao@google.com>
Reviewed-by: Amirali Abdolrashidi <abdolrashidi@google.com>
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Auto-Submit: Charlie Lao <cclao@google.com>