Translator: Give gl_DepthRange its own qualifier
gl_DepthRange and gl_NumSamples are considered "uniform" by the GLSL
spec, but gl_NumSamples gets its own qualifier in the translator.  This
left gl_DepthRange as the only built-in that had to be detected by its
name instead of qualifier.  This change adds a qualifier for
gl_DepthRange to remove this special-ness.

Bug: angleproject:349994211
Change-Id: I2b7b0acf3dcaef22bb4a98e4d35fd2cdfd5c825a
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6998990
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>
Reviewed-by: Yuxin Hu <yuxinhu@google.com>