WGSL: fix do-while when body has continue
If the body of the do while had a "continue" it would
skip over the evaluation of the loop condition and
potentially loop forever.

Change to use the special WGSL "continuing" statement,
which goes at the end of a loop body and always executes,
even if there is an earlier "continue".

Bug: angleproject:42267100
Change-Id: I4ac73e6abcb12e0ff395b83dc5666ac1870724e9
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7003772
Reviewed-by: Liza Burakova <liza@chromium.org>
Commit-Queue: Matthew Denton <mpdenton@chromium.org>
Reviewed-by: Geoff Lang <geofflang@chromium.org>