Translator: Remove NV_shader_framebuffer_fetch
This extension is never exposed by ANGLE.  AFAICT, it was removed in
https://chromium-review.googlesource.com/c/angle/angle/+/688099 and
never added again.

The current implementation is incorrect even if the extension is
exposed; we assume gl_LastFragColorARM is the name, while this extension
uses gl_LastFragColor.  This change removes mentions of this extension
instead of trying to fix it.

Bug: angleproject:349994211
Change-Id: I2391a225db95d73579fd71cf261360973e5935f7
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7004390
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>