Vulkan: Remove mBindingDirtyBitsRequiresPipelineUpdate
mBindingDirtyBitsRequiresPipelineUpdate and
mAttribDirtyBitsRequiresPipelineUpdate are used to avoid
GraphicsPipelineDesc::updateVertexInput() call. This function is not
being used when supportsVertexInputDynamicState is enabled, which almost
all recent drivers do. We could potentially do similar optimizations
when supportsVertexInputDynamicState is enabled to avoid
RenderPassCommands::setVertexInput() call. But the logic is complex
enough not really worth it (See crrev.com/c/6961186 for draft CL). If
any, simply compare the new value and old value probably is as good as
checking the dirty bits. In this CL, all these dirty bits checking are
removed so that we do not waste CPU cycles to track these dirty bits
while not being used at all when supportsVertexInputDynamicState.enabled
is true.

Bug: b/439073246
Bug: b/442636174
Change-Id: I7d71d4fc9388612e7c6ccc50a2e781325fe953bc
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7007241
Commit-Queue: Charlie Lao <cclao@google.com>
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Reviewed-by: Amirali Abdolrashidi <abdolrashidi@google.com>