Clamp up to base level in Metal backend
The Metal backend will crash (i.e., fail a Metal assertion) when
mipmap base level is greater than max level, which can result in
attempting to create a texture view with a range of 0. This change
clamps the max level up to the base level to prevent the crash.

Added tests to verify framebuffer completeness and rendering behavior
when base > max. The tests verify that the framebuffer is complete
when attachment equals base level, and incomplete otherwise.

Test Texture2DBaseMaxTestES3.BaseExceedsMaxFboAttachAtBase is
skipped on Vulkan and OpenGL backends as it currently crashes.

Bug: angleproject:451876199
Change-Id: Ic58668e45a93aa8af05e824d152a411723dfc4cf
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7040665
Commit-Queue: Kenneth Russell <kbr@chromium.org>
Reviewed-by: Kimmo Kinnunen <kkinnunen@apple.com>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>