Translator: Validate duplicate cases earlier
Instead of waiting for the whole switch node to finish and validate it
(with the pass in ValidateSwitch.cpp), this change prepares to validate
everything about switches as soon as possible.

In this change, duplicate case/default values are moved out of
ValidateSwitch and done as soon as the case labels are encountered.

Bug: angleproject:349994211
Change-Id: If6bc4fe7c957d13afd717b6ba5ac419dce7f7827
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7042862
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>
Reviewed-by: Yuxin Hu <yuxinhu@google.com>
Reviewed-by: Geoff Lang <geofflang@chromium.org>