Vulkan: Limit atomic counters to 32
Technically, the limit for atomic counters can be very large; we use an
SSBO to emulate them, which can commonly be gigabytes.

However, some tests create as many atomic counters as possible, which
with the added logic to use those atomic counters result in huge
shaders, risking exhausting SPIR-V's 16-bit ids.

Given most GLES drivers expose a limit of 8, there's no reason for ANGLE
to expose 4096 of them.  This limit is lowered to 32 in this change.

Bug: angleproject:42262227
Change-Id: I9f1db94bb688050cbd9b4eaa7af65317ab806322
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7042357
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>
Reviewed-by: Charlie Lao <cclao@google.com>