Ensure the address that GLshort is written to is 2 bytes aligned
Add an assert to check the address we write GLshort data to is aligned
to 2 bytes. If the assert ever fails, we should add padding to make
sure the address is 2 byte aligned so that there is no extra memory
fetch when we write 2 byte data.

Bug: angleproject:405795981
Change-Id: Ia16b89f20f0160171b7c168e19c660c373f7ae23
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7049344
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Commit-Queue: Yuxin Hu <yuxinhu@google.com>