Vulkan: Fix bytesToAllocate calculation overflow bug
VertexArrayVk::updateStreamedAttribs() is calculating bytesToAllocate =
instanceCount * stride. The problem is that instanceCount and stride are
uint32 but bytesToAllocate is uint64. When the result of multiplication
exceeds UINT_MAX, it becomes zero which causes DynamicBuffer to crash.
This CL makes the calculation done in 64 bit math so that it will not
overflow. The buffer allocation code will fail when size exceeds its
limit and should handle in proper way.

Also Added a test that mimic webGL test with robust buffer access
enabled to demonstrate the crash in ANGLE vulkan backend.

Bug: angleproject:435228814
Change-Id: I10c9aa124fe210dfb1d630fd14493ce03506f368
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7012273
Commit-Queue: Charlie Lao <cclao@google.com>
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Reviewed-by: Geoff Lang <geofflang@chromium.org>