Properly clean up contexts in multi-threading tests
When the context is not unmade-current, it stays alive in the thread
even if deleted.  Seemingly, it also does not get destroyed if the
thread is terminated (a leak?).  As a side effect, the context's
compiler is never freed and sh::Initialize and sh::Finalize go out of
balance.

Bug: angleproject:349994211
Change-Id: Iba0075ae32ef5b956b7ade400130604a2e5776c5
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7052462
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>
Reviewed-by: Amirali Abdolrashidi <abdolrashidi@google.com>