Translator: Fix gl_in redeclaration more
In case gl_in was declared before the input primitive layout was
specified, the shader-declared gl_in was lost.  Later, the symbol-table
would create a gl_in variable with the correct array size, and then the
shader's gl_in declaration was replaced with the symbol table's.

This is incorrect, as the shader-declared gl_in may have different
precision or members.  Instead, this change makes sure that the symbol
table remembers the redeclared gl_in, even if not yet sized, and does
not declare its own.

Bug: angleproject:42264111
Bug: angleproject:349994211
Bug: angleproject:447556589
Change-Id: I3b6b76be8d91fa4bb25c07b2ff185bfb779bc469
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7052463
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>
Reviewed-by: Yuxin Hu <yuxinhu@google.com>