WGSL: Implement out/inout params with pointers.
Out/inout parameters are converted to pointers in WGSL.
WGSL's pointer aliasing rules make this quite difficult,
so all calls to functions that use outparams are converted
into calls to a new wrapper function that use individual
temporaries. Using individual temporaries makes it impossible
for the pointers to alias.

Outparams are required to be l-values in GLSL, but even
l-values can have side effects (x[i++] = ...). This tells
the WGSL translator to take a pointer to these l-values
and store it so the l-value expression is only evaluated
once. Then only the pointer is used after that.

Bug: angleproject:42267100
Change-Id: Ic639df0fe280b95fbc1a9b7ad146c23dc28035de
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7057453
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Commit-Queue: Matthew Denton <mpdenton@chromium.org>