Trace/Replay: Use flush before end GPU timestamp on Native GLES
It was observed that on Adreno 750 the query was inserted before
previous draw commands if there is no flush right before the
glQueryCounterEXT() call. As the result, measured GPU time was 11.449 ms
instead of 15.722 ms in a 100% GPU-bound test with Wall time 15.748 ms.

Similar problem happens in many traces on Mali-G78 but in this case the
flush call is ignored. Using glFenceSync() call achieves the desired
effect. While this fixed the problem in Window mode, in most Offscreen
tests the problem partially remains. Adding glFinish() before the start
GPU timestamp fixed the problem, but can't be used on practice. However,
it seems that problem is also fixed by using the
`-add-swap-into-gpu-time` option.

Bug: angleproject:454646311
Change-Id: I1e94b334d826e723d20d14b4a92b8ef3e3c0fe9a
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7080031
Reviewed-by: Cody Northrop <cnorthrop@google.com>
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Commit-Queue: Igor Nazarov <i.nazarov@samsung.com>